fix(mcp): use int64 for proposals_pending, dropping the int64 to int narrowing
The proposed area/goal/project counts come from count(*) (int64); summing them
into an int field required an int64->int conversion that narrows on 32-bit
builds. The counts are far too small to overflow in practice, but typing the
field as int64 matches the source and removes the conversion entirely rather
than guarding an impossible case. No behaviour or wire change — JSON still
serialises a bare number.

Addresses Augment review on PR #20.

+++ internal/mcp/brief.go
@@ -103,7 +103,7 @@ type BriefOutput struct {
 	ActiveGoals      []goal.ActiveGoalSummary `json:"-"`
 	RSSHighlights    []RSSHighlight           `json:"-"`
 	ContentPipeline  []ContentSummary         `json:"-"`
-	ProposalsPending int                      `json:"-"`
+	ProposalsPending int64                    `json:"-"`
 
 	// Reflection fields.
 	PlannedItems   []daily.Item `json:"-"`
@@ -130,8 +130,9 @@ type briefMorningWire struct {
 	// ProposalsPending is the summed count of agent-proposed area/goal/project
 	// drafts awaiting owner triage. Unlike the list fields it is a scalar, so
 	// it always serialises (0 when nothing is pending) — the push consumer
-	// gates its nudge on N > 0.
-	ProposalsPending int `json:"proposals_pending"`
+	// gates its nudge on N > 0. int64 matches the count(*) source type, so no
+	// narrowing conversion is needed.
+	ProposalsPending int64 `json:"proposals_pending"`
 }
 
 // briefReflectionWire is the wire shape for mode=reflection. Field tags mirror
@@ -425,7 +426,7 @@ func (s *Server) fillProposalsPending(ctx context.Context, out *BriefOutput) {
 	} else {
 		s.logger.Warn("brief: proposed projects count", "error", err)
 	}
-	out.ProposalsPending = int(total)
+	out.ProposalsPending = total
 }
 
 // fillBriefReflection populates the reflection fields: the day's plan items
